code cleanup in progress
